Added getLastWeaponID and replaceLastWithID to weapon utils.
This allows for 'last' to be used in commands and it will be replaced
with the last weapon the user has found.

where getLastWeaponID is the actually query / finding of the ID and
replaceLastWithID just loops through the args checking if its either
last or l and replaces it with getLastWeaponID.

Intent: QOL improvement, since typing weapon codes after opening a
crate on mobile can be quite annoying. owo crate + owo w last is easier.